kickstarter/requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
cachetools==5.0.0
category-encoders==2.3.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
dnspython==2.2.0
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
Flask==2.0.2
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.8.0
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
protobuf==3.19.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
stack-data==0.1.4
statsmodels==0.13.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.11.4
tomli==2.0.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
